+++ lib/src/annotation/circle_annotation_manager.dart
@@ -37,11 +37,11 @@ class CircleAnnotationManager extends BaseAnnotationManager {
   /// Delete all the annotation added by this manager.
   Future<void> deleteAll() => messenger.deleteAll(id);
 
-  /// Controls the intensity of light emitted on the source features. This property works only with 3D light, i.e. when `lights` root property is defined.
+  /// Controls the intensity of light emitted on the source features.
   Future<void> setCircleEmissiveStrength(double circleEmissiveStrength) =>
       messenger.setCircleEmissiveStrength(id, circleEmissiveStrength);
 
-  /// Controls the intensity of light emitted on the source features. This property works only with 3D light, i.e. when `lights` root property is defined.
+  /// Controls the intensity of light emitted on the source features.
   Future<double?> getCircleEmissiveStrength() =>
       messenger.getCircleEmissiveStrength(id);