+++ lib/src/style/layer/raster_layer.dart
@@ -11,12 +11,15 @@ class RasterLayer extends Layer {
     slot,
     required this.sourceId,
     this.sourceLayer,
+    this.rasterArrayBand,
     this.rasterBrightnessMax,
     this.rasterBrightnessMin,
     this.rasterColor,
     this.rasterColorMix,
     this.rasterColorRange,
     this.rasterContrast,
+    this.rasterElevation,
+    this.rasterEmissiveStrength,
     this.rasterFadeDuration,
     this.rasterHueRotate,
     this.rasterOpacity,
@@ -38,13 +41,16 @@ class RasterLayer extends Layer {
   /// A source layer is an individual layer of data within a vector source. A vector source can have multiple source layers.
   String? sourceLayer;
 
+  /// Displayed band of raster array source layer
+  String? rasterArrayBand;
+
   /// Increase or reduce the brightness of the image. The value is the maximum brightness.
   double? rasterBrightnessMax;
 
   /// Increase or reduce the brightness of the image. The value is the minimum brightness.
   double? rasterBrightnessMin;
 
-  /// Defines a color map by which to colorize a raster layer, parameterized by the `["raster-value"]` expression and evaluated at 1024 uniformly spaced steps over the range specified by `raster-color-range`.
+  /// Defines a color map by which to colorize a raster layer, parameterized by the `["raster-value"]` expression and evaluated at 256 uniformly spaced steps over the range specified by `raster-color-range`.
   int? rasterColor;
 
   /// When `raster-color` is active, specifies the combination of source RGB channels used to compute the raster value. Computed using the equation `mix.r * src.r + mix.g * src.g + mix.b * src.b + mix.a`. The first three components specify the mix of source red, green, and blue channels, respectively. The fourth component serves as a constant offset and is *not* multipled by source alpha. Source alpha is instead carried through and applied as opacity to the colorized result. Default value corresponds to RGB luminosity.
@@ -55,6 +61,12 @@ class RasterLayer extends Layer {
 
   /// Increase or reduce the contrast of the image.
   double? rasterContrast;
+
+  /// Specifies an uniform elevation from the ground, in meters. Only supported with image sources.
+  double? rasterElevation;
+
+  /// Controls the intensity of light emitted on the source features.
+  double? rasterEmissiveStrength;
 
   /// Fade duration when a new tile is added.
   double? rasterFadeDuration;
@@ -79,6 +91,9 @@ class RasterLayer extends Layer {
           visibility?.toString().split('.').last.toLowerCase();
     }
     var paint = {};
+    if (rasterArrayBand != null) {
+      paint["raster-array-band"] = rasterArrayBand;
+    }
     if (rasterBrightnessMax != null) {
       paint["raster-brightness-max"] = rasterBrightnessMax;
     }
@@ -96,6 +111,12 @@ class RasterLayer extends Layer {
     }
     if (rasterContrast != null) {
       paint["raster-contrast"] = rasterContrast;
+    }
+    if (rasterElevation != null) {
+      paint["raster-elevation"] = rasterElevation;
+    }
+    if (rasterEmissiveStrength != null) {
+      paint["raster-emissive-strength"] = rasterEmissiveStrength;
     }
     if (rasterFadeDuration != null) {
       paint["raster-fade-duration"] = rasterFadeDuration;
@@ -159,6 +180,9 @@ class RasterLayer extends Layer {
               .last
               .toLowerCase()
               .contains(map["layout"]["visibility"])),
+      rasterArrayBand: map["paint"]["raster-array-band"] is String?
+          ? map["paint"]["raster-array-band"] as String?
+          : null,
       rasterBrightnessMax: map["paint"]["raster-brightness-max"] is num?
           ? (map["paint"]["raster-brightness-max"] as num?)?.toDouble()
           : null,
@@ -174,6 +198,12 @@ class RasterLayer extends Layer {
           .toList(),
       rasterContrast: map["paint"]["raster-contrast"] is num?
           ? (map["paint"]["raster-contrast"] as num?)?.toDouble()
+          : null,
+      rasterElevation: map["paint"]["raster-elevation"] is num?
+          ? (map["paint"]["raster-elevation"] as num?)?.toDouble()
+          : null,
+      rasterEmissiveStrength: map["paint"]["raster-emissive-strength"] is num?
+          ? (map["paint"]["raster-emissive-strength"] as num?)?.toDouble()
           : null,
       rasterFadeDuration: map["paint"]["raster-fade-duration"] is num?
           ? (map["paint"]["raster-fade-duration"] as num?)?.toDouble()